Fixes a thread race when cancelling an async call across threads
Note that this also removes the possibility to fully abort a cancellation
in a synchronizer-wrapped method. You can still catch and handle cancellations
in the wrapped code, and the await in the caller thread will not resolve until
the cancellation completes, but with this change the calling thread will always
receive a CancelledError even if the called code doesn't reraise it.