Fix another very subtle cancellation race
There is a risk of a race inside asyncio.wait_for between receiving a CancellationError
and checking fut.done(), which can cause well behaving cancellation to get swallowed.
This adds an additional outer asyncio.shield layer to prevent this.

Basically impossible to create a repro without monkeypatching asyncio.wait_for which
kind of defeats the purpose